change relativeHumidity type in Weather entity from Double to Int for consistency with API response

+++ src/main/kotlin/domain/entity/Weather.kt
@@ -6,7 +6,7 @@ import org.beijingteam.domain.entity.enums.WeatherCondition
 
 data class Weather(
     val temperature: Double,
-    val relativeHumidity: Double,
+    val relativeHumidity: Int,
     val weatherCondition: WeatherCondition,
     val temperatureCategory: TemperatureCategory
 )
